Show favorites group
Previous commit, while useful, didn't actually resolve it. Testing this
thing is so troublesome, I guess maybe I was seeing the live version
thinking it was the test. Because the UIMenu dom isn't attached when the
UIFavoritesGroup() constructor is called it can't find the
#starred-emotes-group. This changes it to look for that under the
UIMenu's dom, which will find it even while unattached.